Rename to radiationCRAX and add binary for macOS
