+++ cea/default.config
@@ -125,8 +125,7 @@ weather =
 weather.type = WeatherPathParameter
 weather.help = Path to a weather file (.epw) or the label of one CEA Built-in weather file. Leave blank to fetch from third-party sources.
 
-
-[radiation-usr]
+[radiation-crax]
 buildings =
 buildings.type = BuildingsParameter
 buildings.help = List of buildings considered for the radiation simulation in the format of [BuildingID1, BuildingID2, ...]. Leave blank when simulating all buildings.
@@ -134,7 +133,7 @@ buildings.help = List of buildings considered for the radiation simulation in th
 albedo = 0.2
 albedo.type = RealParameter
 albedo.help = Albedo of the terrain.
-albedo.category = Parameters for usr
+albedo.category = Parameters for crax
 
 calculate-sensor-data = true
 calculate-sensor-data.type = BooleanParameter
@@ -143,23 +142,38 @@ calculate-sensor-data.category = Advanced
 
 using-cea-sensor = true
 using-cea-sensor.type = BooleanParameter
-using-cea-sensor.help =  True to use the cea-daysim method to generate sensor csv files. False to use the USR method to generate sensor csv files (more fast).
+using-cea-sensor.help =  True to use the cea-daysim method to generate sensor csv files. False to use the CRAX method to generate sensor csv files (more fast).
 using-cea-sensor.category = Advanced
 
-roof-grid = 1
+roof-grid = 3
 roof-grid.type = RealParameter
-roof-grid.help =  Grid resolution of the roof surfaces (m) for USR grid generation method.
+roof-grid.help =  Grid resolution of the roof surfaces (m) for CRAX grid generation method.
 roof-grid.category = Level of Details
 
-walls-grid = 1
+walls-grid = 3
 walls-grid.type = RealParameter
-walls-grid.help = Grid resolution of wall surfaces (m) for USR grid generation method.
+walls-grid.help = Grid resolution of wall surfaces (m) for CRAX grid generation method.
 walls-grid.category = Level of Details
 
-usr-bin-directory =
-usr-bin-directory.type = PathParameter
-usr-bin-directory.help = Path to the USR binaries. This path will be used instead of the default CEA Dependencies.
-usr-bin-directory.category = Advanced
+zone-geometry = 2
+zone-geometry.type = RealParameter
+zone-geometry.help = Simplification level of zone geometry; tolerance distance of points in meters from original.
+zone-geometry.category = Level of Details
+
+surrounding-geometry = 5
+surrounding-geometry.type = RealParameter
+surrounding-geometry.help = Simplification level of surroundings geometry; tolerance distance of points in meters from original.
+surrounding-geometry.category = Level of Details
+
+neglect-adjacent-buildings = false
+neglect-adjacent-buildings.type = BooleanParameter
+neglect-adjacent-buildings.help = True if adjacent walls with neighboring buildings should be neglected. If set to True, the results might be less accurate (there might be insolation on attached walls and no adiabatic surfaces between adjacent buildings will be considered), but the simulation time will decrease somewhat.
+neglect-adjacent-buildings.category = Level of Details
+
+crax-bin-directory =
+crax-bin-directory.type = PathParameter
+crax-bin-directory.help = Path to the CRAX binaries. This path will be used instead of the default CEA Dependencies.
+crax-bin-directory.category = Advanced
 
 update-shadow-day = 15
 update-shadow-day.type = RealParameter
